Struct arcadeum::crypto::MerkleTree

pub struct MerkleTree<T: MerkleLeaf> { /* fields omitted */ }

Balanced Merkle tree

The leaves of the tree are positioned according to their indices in truncated binary encoding. This results in every Merkle proof having a number of hashes within one of any other proof, even when the number of elements isn't a power of two.

Examples

let tree = arcadeum::crypto::MerkleTree::with_salt(
    vec![
        vec![0; 0],
        vec![1; 1],
        vec![2; 2],
        vec![3; 3],
        vec![4; 4],
        vec![5; 5],
        vec![6; 6],
        vec![7; 7],
        vec![8; 8],
        vec![9; 9],
    ],
    16,
    &mut rand::thread_rng(),
)
.unwrap();

assert_eq!(
    arcadeum::crypto::MerkleTree::deserialize(&tree.serialize()).unwrap(),
    tree
);

for i in 0..tree.len() {
    let proof = tree.proof(i).unwrap();

    assert_eq!(proof.root(), tree.root());

    assert_eq!(
        arcadeum::crypto::MerkleProof::deserialize(&proof.serialize()).unwrap(),
        proof
    );
}

Implementations

impl<T: MerkleLeaf> MerkleTree<T>

pub fn new(elements: Vec<T>) -> Self

Constructs an unsalted Merkle tree from a vector.

See MerkleTree::with_salt.

pub fn with_salt(
elements: Vec<T>,
salt_size: usize,
random: &mut dyn RngCore
) -> Result<Self, String>

Constructs a salted Merkle tree from a vector and a source of entropy.

See MerkleTree::new.

pub fn deserialize(data: &[u8]) -> Result<Self, String>

Constructs a Merkle tree from its binary representation.

data must have been constructed using MerkleTree::serialize.

pub fn serialize(&self) -> Vec<u8>

Generates a binary representation that can be used to reconstruct the Merkle tree.

See MerkleTree::deserialize.

pub fn elements(&self) -> &Vec<T>

Gets the elements in the Merkle tree.

pub fn is_empty(&self) -> bool

true if the Merkle tree is empty.

pub fn len(&self) -> usize

Gets the number of elements in the Merkle tree.

pub fn root(&self) -> &Hash

Gets the root hash of the Merkle tree.

pub fn proof(&self, index: usize) -> Result<MerkleProof<T>, String>

Generates a Merkle proof for the element at the given index.