Struct arcadeum::store::Context

pub struct Context<S: State> { /* fields omitted */ }

State::apply utilities

Implementations

impl<S: State> Context<S>

pub fn mutate_secret(
&mut self,
player: Player,
mutate: impl Fn(&mut S::Secret, &mut dyn RngCore, &mut dyn FnMut(S::Event))
)

Mutates a player's secret information.

pub async fn reveal<T: Secret, '_>(
&'_ mut self,
player: Player,
reveal: impl Fn(&S::Secret) -> T + 'static,
verify: impl Fn(&T) -> bool + 'static
) -> T

Requests a player's secret information.

The random number generator is re-seeded after this call to prevent players from influencing the randomness of the state via trial and error.

See Context::reveal_unique for a faster non-re-seeding version of this method.

pub async fn reveal_unique<T: Secret, '_>(
&'_ mut self,
player: Player,
reveal: impl Fn(&S::Secret) -> T + 'static,
verify: impl Fn(&T) -> bool + 'static
) -> T

Requests a player's secret information.

The random number generator is not re-seeded after this call, so care must be taken to guarantee that the verify function accepts only one possible input. Without this guarantee, players can influence the randomness of the state via trial and error.

See Context::reveal for a slower re-seeding version of this method.

pub fn random(&mut self) -> impl Future<Output = impl RngCore>

Constructs a random number generator via commit-reveal.

pub fn log(&mut self, event: S::Event)

Logs an event if logging is enabled.

See Context::enable_logs.

pub fn enable_logs(&mut self, enabled: bool)

Enables or disables logging.

See Context::log.