Trait arcadeum::store::State

pub trait State: Clone {
    type ID: ID;
    type Nonce: Nonce;
    type Action: Action + Debug;
    type Secret: Secret;
    fn version() -> &'static [u8];

    fn deserialize(data: &[u8]) -> Result<Self, String>;

    fn serialize(&self) -> Option<Vec<u8>>;

    fn verify(
        &self, 
        player: Option<Player>, 
        action: &Self::Action
    ) -> Result<(), String>;

    fn apply(
        self, 
        player: Option<Player>, 
        action: &Self::Action, 
        context: Context<Self>
    ) -> Pin<Box<dyn Future<Output = (Self, Context<Self>)>>>;

    fn certificate(address: &Address) -> String { ... }

    fn is_serializable(&self) -> bool { ... }
}

Domain-specific store state trait

Associated Types

type ID: ID

Identifier type

type Nonce: Nonce

Nonce type

type Action: Action + Debug

Action type

type Secret: Secret

Secret type

Loading content...
Required methods

fn version() -> &'static [u8]

Gets the ABI version of this implementation.

See super::tag and super::version::version for potentially helpful utilities.

fn deserialize(data: &[u8]) -> Result<Self, String>

Constructs a state from its binary representation.

data must have been constructed using State::serialize.

fn serialize(&self) -> Option<Vec<u8>>

Generates a binary representation that can be used to reconstruct the state.

See State::deserialize and State::is_serializable.

fn verify(
&self,
player: Option<Player>,
action: &Self::Action
) -> Result<(), String>

Verifies if an action by a given player is valid for the state.

fn apply(
self,
player: Option<Player>,
action: &Self::Action,
context: Context<Self>
) -> Pin<Box<dyn Future<Output = (Self, Context<Self>)>>>

Applies an action by a given player to the state.

Loading content...
Provided methods

fn certificate(address: &Address) -> String

Formats the message that must be signed in order to certify the subkey for a given address.

fn is_serializable(&self) -> bool

Checks if the state has a binary representation.

This should be implemented whenever possible for improved performance. The return value must agree with State::serialize.

Loading content...