Trait arcadeum::State

pub trait State: Clone {
    type ID: ID;
    type Nonce: Nonce;
    type Action: Action;
    pub fn version() -> &'static [u8];

    pub fn deserialize(data: &[u8]) -> Result<Self, String>;

    pub fn serialize(&self) -> Option<Vec<u8>>;

    pub fn apply(
        &mut self, 
        player: Option<Player>, 
        action: &Self::Action
    ) -> Result<(), String>;

    pub fn challenge(address: &Address) -> String { ... }

    pub fn approval(player: &Address, subkey: &Address) -> String { ... }

    pub fn is_serializable(&self) -> bool { ... }
}

Domain-specific state trait

Associated Types

type ID: ID

Identifier type

type Nonce: Nonce

Nonce type

type Action: Action

Action type

Loading content...
Required methods

pub fn version() -> &'static [u8]

Gets the ABI version of this implementation.

See tag and version::version for potentially helpful utilities.

pub fn deserialize(data: &[u8]) -> Result<Self, String>

Constructs a state from its binary representation.

data must have been constructed using State::serialize.

pub fn serialize(&self) -> Option<Vec<u8>>

Generates a binary representation that can be used to reconstruct the state.

See State::deserialize and State::is_serializable.

pub fn apply(
&mut self,
player: Option<Player>,
action: &Self::Action
) -> Result<(), String>

Applies an action by a given player to the state.

Loading content...
Provided methods

pub fn challenge(address: &Address) -> String

Gets the challenge that must be signed in order to certify the subkey with the given address.

pub fn approval(player: &Address, subkey: &Address) -> String

Gets the approval that must be signed by the owner in order to approve a subkey for a player.

pub fn is_serializable(&self) -> bool

Checks if the state has a binary representation.

This should be implemented whenever possible for improved performance. The return value must agree with State::serialize.

Loading content...