Function arcadeum::version::tag

pub fn tag<P: AsRef<Path>>(
    module: impl AsRef<Path>, 
    identifier: &str, 
    paths: impl Iterator<Item = P>
) -> Result<()>

Generates a module exporting a canonical digest of the contents of the files and directories at the given paths.

The module is written to the given module path. The version is exported as a constant with the given identifier.

To use the generated code, it must be include!d in your code:

include!(concat!(env!("OUT_DIR"), "/generated.rs"));

The traversal of paths is deterministic. Directories are traversed recursively.

Examples

let mut module = std::path::PathBuf::new();
module.push(std::env::var("OUT_DIR").unwrap());
module.push("generated.rs");

arcadeum::tag(
    module,
    "VERSION",
    ["Cargo.toml", "Cargo.lock", "build.rs", "src"].iter(),
)
.unwrap();